fix: Uses pull query metrics for all paths, not just /query
