+++ ksqldb-rest-app/src/main/java/io/confluent/ksql/rest/server/execution/PullQueryExecutor.java
@@ -110,6 +110,7 @@ import java.util.Set;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+import org.apache.kafka.common.utils.Time;
 import org.apache.kafka.connect.data.Field;
 import org.apache.kafka.connect.data.Struct;
 import org.slf4j.Logger;
@@ -139,17 +140,22 @@ public final class PullQueryExecutor {
   private final KsqlExecutionContext executionContext;
   private final RoutingFilterFactory routingFilterFactory;
   private final RateLimiter rateLimiter;
+  private final Optional<PullQueryExecutorMetrics> pullQueryMetrics;
+  private final Time time;
 
   public PullQueryExecutor(
       final KsqlExecutionContext executionContext,
       final RoutingFilterFactory routingFilterFactory,
-      final KsqlConfig ksqlConfig
+      final KsqlConfig ksqlConfig,
+      final Optional<PullQueryExecutorMetrics> pullQueryMetrics
   ) {
     this.executionContext = Objects.requireNonNull(executionContext, "executionContext");
     this.routingFilterFactory =
         Objects.requireNonNull(routingFilterFactory, "routingFilterFactory");
     this.rateLimiter = RateLimiter.create(ksqlConfig.getInt(
         KsqlConfig.KSQL_QUERY_PULL_MAX_QPS_CONFIG));
+    this.pullQueryMetrics = pullQueryMetrics;
+    this.time = Time.SYSTEM;
   }
 
   @SuppressWarnings("unused") // Needs to match validator API.
@@ -165,7 +171,6 @@ public final class PullQueryExecutor {
   public PullQueryResult execute(
       final ConfiguredStatement<Query> statement,
       final ServiceContext serviceContext,
-      final Optional<PullQueryExecutorMetrics> pullQueryMetrics,
       final Optional<Boolean> isInternalRequest
   ) {
     if (!statement.getStatement().isPullQuery()) {
@@ -182,6 +187,7 @@ public final class PullQueryExecutor {
     }
 
     try {
+      final long startTime = time.nanoseconds();
       final RoutingOptions routingOptions = new ConfigRoutingOptions(
           statement.getConfig(), statement.getConfigOverrides(), statement.getRequestProperties());
       // If internal listeners are in use, we require the request to come from that listener to
@@ -223,13 +229,21 @@ public final class PullQueryExecutor {
           contextStacker,
           pullQueryMetrics);
 
-      return handlePullQuery(
+      PullQueryResult result = handlePullQuery(
           statement,
           executionContext,
           serviceContext,
           pullQueryContext,
           routingOptions
       );
+
+      if (pullQueryMetrics.isPresent()) {
+        //Record latency at microsecond scale
+        final double latency = (time.nanoseconds() - startTime) / 1000f;
+        pullQueryMetrics.get().recordLatency(latency);
+        pullQueryMetrics.get().recordRate(1);
+      }
+      return result;
     } catch (final Exception e) {
       pullQueryMetrics.ifPresent(metrics -> metrics.recordErrorRate(1));
       throw new KsqlStatementException(
@@ -245,6 +259,12 @@ public final class PullQueryExecutor {
     if (!rateLimiter.tryAcquire()) {
       throw new KsqlException("Host is at rate limit for pull queries. Currently set to "
           + rateLimiter.getRate() + " qps.");
+    }
+  }
+
+  public void closeMetrics() {
+    if (pullQueryMetrics != null) {
+      pullQueryMetrics.ifPresent(PullQueryExecutorMetrics::close);
     }
   }
