+++ ksqldb-rest-app/src/test/java/io/confluent/ksql/rest/server/resources/streaming/PullQueryPublisherTest.java
@@ -83,7 +83,7 @@ public class PullQueryPublisherTest {
         pullQueryExecutor);
 
     PullQueryResult result = new PullQueryResult(entity, Optional.empty());
-    when(pullQueryExecutor.execute(any(), any(), any(), any())).thenReturn(result);
+    when(pullQueryExecutor.execute(any(), any(), any())).thenReturn(result);
     when(entity.getSchema()).thenReturn(SCHEMA);
 
     doAnswer(callRequestAgain()).when(subscriber).onNext(any());
@@ -107,8 +107,7 @@ public class PullQueryPublisherTest {
     subscription.request(1);
 
     // Then:
-    verify(pullQueryExecutor).execute(statement, serviceContext, Optional.empty(),
-        Optional.of(false));
+    verify(pullQueryExecutor).execute(statement, serviceContext, Optional.of(false));
   }
 
   @Test
@@ -121,8 +120,7 @@ public class PullQueryPublisherTest {
 
     // Then:
     verify(subscriber).onNext(any());
-    verify(pullQueryExecutor).execute(statement, serviceContext, Optional.empty(),
-        Optional.of(false));
+    verify(pullQueryExecutor).execute(statement, serviceContext, Optional.of(false));
   }
 
   @Test
@@ -157,7 +155,7 @@ public class PullQueryPublisherTest {
     // Given:
     givenSubscribed();
     final Throwable e = new RuntimeException("Boom!");
-    when(pullQueryExecutor.execute(any(), any(), any(), any())).thenThrow(e);
+    when(pullQueryExecutor.execute(any(), any(), any())).thenThrow(e);
 
     // When:
     subscription.request(1);
